Generate Overdispersed (SSI) Point Pattern for Landscape Sites

This function is a low-level utility to generate spatial point patterns for the MICRO-LANDSCAPE point sets.

pointsOverdispersed(n, inhibitionFactor = 1, xLim = c(0, 1), yLim = c(0,
  1))

Arguments

n
number of points to generate

inhibitionFactor
controls level of overdispersion (higher values correspond to a more overdispersed spatial point process)

xLim
a length 2 numeric vector of bounds for the sampling grid

yLim
a length 2 numeric vector of bounds for the sampling grid

Value

a list with two elements x and y corresponding to the sampled points

Examples

pointsOverdispersed(n=10, inhibitionFactor = 1, xLim=c(0,1), yLim=c(0,1))

#> Warning: Gave up after 1000 attempts with only 9 points placed out of 10

#> Warning: Gave up after 1000 attempts with only 9 points placed out of 10

#> $x
#>  [1] 0.81891544 0.48114785 0.11929568 0.98754066 0.04333774 0.61551735
#>  [7] 0.37428498 0.93487477 0.02417255 0.84878830
#> 
#> $y
#>  [1] 0.99964589 0.90675787 0.51892113 0.32332248 0.18201160 0.37299683
#>  [7] 0.00174157 0.69702311 0.85376548 0.03470692
#>